ENH: fixed bug #164

+++ core/controllers/ItemController.php
@@ -97,15 +97,16 @@ class ItemController extends AppController
       {
       $deleteMetadata = $this->_getParam('deleteMetadata');
       $editMetadata = $this->_getParam('editMetadata');
-      if(isset($deleteMetadata) && $this->view->isModerator) //remove users from group
+      if(isset($deleteMetadata) && !empty($deleteMetadata) && $this->view->isModerator) //remove users from group
         {
         $this->disableView();
         $this->disableLayout();
         $metadataId = $this->_getParam('element');
         $this->ItemRevision->deleteMetadata($itemRevision, $metadataId);
         echo JsonComponent::encode(array(true, $this->t('Changes saved')));
+        return;
         }
-      if(isset($editMetadata) && $this->view->isModerator) //remove users from group
+      if(isset($editMetadata)&& !empty($editMetadata) && $this->view->isModerator) //remove users from group
         {
         $metadataId = $this->_getParam('metadataId');
         $metadatatype = $this->_getParam('metadatatype');
